Refactored creating the files and covered these with tests
