--- internal/cli/helper/gofile/functions.go
@@ -1,8 +1,0 @@
-package cli
-
-import "os/exec"
-
-// FixGoimports fixes the imports in the file
-func FixGoimports(path string) error {
-	return exec.Command("goimports", "-w", path).Run()
-}
