+++ internal/cli/helper/output/files/testdata/create_expected.txt
@@ -1,0 +1,1 @@
+Hello, World!
