+++ internal/cli/helper/output/files/testdata/fix_goimports.txt
@@ -1,0 +1,9 @@
+package testdata
+
+import (
+	"strings"
+)
+
+func doNothing() {
+	return nil
+}
