+++ internal/cli/helper/output/files/testdata/fix_goimports_expected.txt
@@ -1,0 +1,5 @@
+package testdata
+
+func doNothing() {
+	return nil
+}
