+++ internal/cli/make/config/testdata/config_go_expected.txt
